refactor: fix package name user.repository -> user.persistence
