+++ src/main/java/com/meoguri/linkocean/configuration/security/oauth/CustomOAuth2UserService.java
@@ -16,7 +16,7 @@ import org.springframework.stereotype.Service;
 import com.meoguri.linkocean.domain.user.entity.User;
 import com.meoguri.linkocean.domain.user.entity.vo.Email;
 import com.meoguri.linkocean.domain.user.entity.vo.OAuthType;
-import com.meoguri.linkocean.domain.user.repository.UserRepository;
+import com.meoguri.linkocean.domain.user.persistence.UserRepository;
 
 import lombok.extern.slf4j.Slf4j;
