+++ src/main/java/com/meoguri/linkocean/domain/user/persistence/FindUserByIdQuery.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.user.repository;
+package com.meoguri.linkocean.domain.user.persistence;
 
 import static java.lang.String.*;
