+++ src/main/java/com/meoguri/linkocean/domain/user/persistence/UserRepository.java
@@ -1,4 +1,4 @@
-package com.meoguri.linkocean.domain.user.repository;
+package com.meoguri.linkocean.domain.user.persistence;
 
 import java.util.Optional;
