+++ src/test/java/com/meoguri/linkocean/domain/profile/persistence/CheckIsFollowQueryTest.java
@@ -16,7 +16,7 @@ import org.springframework.context.annotation.Import;
 import com.meoguri.linkocean.domain.profile.entity.Follow;
 import com.meoguri.linkocean.domain.profile.entity.Profile;
 import com.meoguri.linkocean.domain.user.entity.User;
-import com.meoguri.linkocean.domain.user.repository.UserRepository;
+import com.meoguri.linkocean.domain.user.persistence.UserRepository;
 
 @Import(CheckIsFollowQuery.class)
 @DataJpaTest
