+++ src/test/java/com/meoguri/linkocean/domain/profile/persistence/CustomProfileRepositoryImplTest.java
@@ -20,7 +20,7 @@ import com.meoguri.linkocean.domain.profile.entity.Follow;
 import com.meoguri.linkocean.domain.profile.entity.Profile;
 import com.meoguri.linkocean.domain.profile.persistence.dto.ProfileFindCond;
 import com.meoguri.linkocean.domain.user.entity.User;
-import com.meoguri.linkocean.domain.user.repository.UserRepository;
+import com.meoguri.linkocean.domain.user.persistence.UserRepository;
 import com.meoguri.linkocean.support.common.CustomP6spySqlFormat;
 
 @Import(CustomP6spySqlFormat.class)
