+++ src/test/java/com/meoguri/linkocean/domain/profile/persistence/FindProfileByIdQueryTest.java
@@ -10,7 +10,7 @@ import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
 import org.springframework.context.annotation.Import;
 
 import com.meoguri.linkocean.domain.profile.entity.Profile;
-import com.meoguri.linkocean.domain.user.repository.UserRepository;
+import com.meoguri.linkocean.domain.user.persistence.UserRepository;
 
 @Import(FindProfileByIdQuery.class)
 @DataJpaTest
