+++ src/test/java/com/meoguri/linkocean/domain/user/service/UserServiceImplTest.java
@@ -12,7 +12,7 @@ import org.springframework.transaction.annotation.Transactional;
 import com.meoguri.linkocean.domain.user.entity.User;
 import com.meoguri.linkocean.domain.user.entity.vo.Email;
 import com.meoguri.linkocean.domain.user.entity.vo.OAuthType;
-import com.meoguri.linkocean.domain.user.repository.UserRepository;
+import com.meoguri.linkocean.domain.user.persistence.UserRepository;
 
 @Transactional
 @SpringBootTest
