+++ src/test/java/com/meoguri/linkocean/support/persistence/BasePersistenceTest.java
@@ -1,0 +1,10 @@
+package com.meoguri.linkocean.support.persistence;
+
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.transaction.annotation.Transactional;
+
+@Transactional
+@DataJpaTest
+public class BasePersistenceTest {
+
+}
